split Dir. SamplingS
