+++ sources/Deltares.Probabilistic/Uncertainty/Direction.h
@@ -1,0 +1,75 @@
+// Copyright (C) Stichting Deltares. All rights reserved.
+//
+// This file is part of the Probabilistic Library.
+//
+// The Probabilistic Library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+//
+// All names, logos, and references to "Deltares" are registered trademarks of
+// Stichting Deltares and remain full property of Stichting Deltares at all times.
+// All rights reserved.
+//
+#pragma once
+#include "DirectionalSamplingSettingsS.h"
+#include "UncertaintyMethod.h"
+
+namespace Deltares::Uncertainty
+{
+    /**
+     * \brief Calculates the sensitivity using the Directional Sampling algorithm
+     */
+    class Result
+    {
+    public:
+        double Distance;
+        double ZValue;
+    };
+
+    class Direction
+    {
+    public:
+        Direction(std::shared_ptr<Models::Sample> sample, int index)
+        {
+            this->sample = sample;
+            this->sample->IterationIndex = index;
+            this->index = index;
+        }
+
+        void AddResult(double distance, double z);
+        double GetDistanceAtZ(double z) const;
+        std::shared_ptr<Models::Sample> CreateNewSampleAt(double z, double maxBeta);
+        bool Valid = true;
+        double lastWeight = 0;
+        double lastDifference = 0;
+        double lastDistance = 0;
+
+        /**
+         * \brief Performs an operation on a direction resulting in a numeric value for a collection of directions
+         * \param directions Collection of samples
+         * \param function Operation on a sample
+         * \return Resulting numeric values
+         */
+        static std::vector<double> select(const std::vector<std::shared_ptr<Direction>>& directions,
+            const std::function<double(const std::shared_ptr<Direction>&)>& function);
+        static std::vector<std::shared_ptr<Direction>> where(std::vector<std::shared_ptr<Direction>>& directions,
+            const std::function<bool(const std::shared_ptr<Direction>&)>& function);
+
+    private:
+        int index = 0;
+        std::shared_ptr<Models::Sample> sample;
+        std::vector<double> distances = std::vector<double>();
+        std::vector<double> zValues = std::vector<double>();
+    };
+
+}
+
