+++ sources/Deltares.Probabilistic/Uncertainty/DirectionalSamplingS.cpp
@@ -325,95 +325,5 @@ namespace Deltares::Uncertainty
         }
     }
 
-    void DirectionalSamplingS::Direction::AddResult(double distance, double z)
-    {
-        distances.push_back(distance);
-        zValues.push_back(z);
-
-        lastDifference = distance - lastDistance;
-        lastDistance = distance;
-        lastWeight = Statistics::StandardNormal::getQFromU(std::abs(distance));
-
-        // Create a list of Result objects
-        std::vector<std::shared_ptr<Result>> results;
-        for (size_t i = 0; i < distances.size(); i++)
-        {
-            std::shared_ptr<Result> tempVar = std::make_shared<Result>();
-            tempVar->Distance = distances[i];
-            tempVar->ZValue = zValues[i];
-            results.push_back(tempVar);
-        }
-
-        // Sort the results based on the Distance property
-        std::sort(results.begin(), results.end(),
-            [](const std::shared_ptr<Result>& p, const std::shared_ptr<Result>& q) {return p->Distance < q->Distance; });
-
-        // Clear the original lists and repopulate them with the sorted values
-        distances.clear();
-        zValues.clear();
-
-        for (const auto& result : results)
-        {
-            distances.push_back(result->Distance);
-            zValues.push_back(result->ZValue);
-        }
-    }
-
-    double DirectionalSamplingS::Direction::GetDistanceAtZ(double z) const
-    {
-        if (Valid)
-        {
-            return NumericSupport::interpolate(z, zValues, distances, true);
-        }
-        else
-        {
-            return 0;
-        }
-    }
-
-    std::shared_ptr<Sample> DirectionalSamplingS::Direction::CreateNewSampleAt(double z, double maxBeta)
-    {
-        double distance = NumericSupport::interpolate(z, zValues, distances, true); //interpolates the distance (d) for the given z value in the direction
-
-        if (!Valid || distance < 0.0 || distance > maxBeta)
-        {
-            if (Valid)
-            {
-                Valid = false;
-            }
-            return sample;
-        }
-        else
-        {
-            std::shared_ptr<Sample> newSample = sample->getSampleAtBeta(distance);
-            newSample->IterationIndex = this->index;
-            return newSample;
-        }
-    }
-
-    std::vector<double> DirectionalSamplingS::Direction::select(std::vector<std::shared_ptr<Direction>>& directions, std::function<double(std::shared_ptr<Direction>)> function)
-    {
-        std::vector<double> result(directions.size());
-        for (size_t i = 0; i < directions.size(); i++)
-        {
-            result[i] = function(directions[i]);
-        }
-        return result;
-    }
-
-    std::vector<std::shared_ptr<DirectionalSamplingS::Direction>> DirectionalSamplingS::Direction::where(std::vector<std::shared_ptr<Direction>>& directions, std::function<bool(std::shared_ptr<Direction>)> function)
-    {
-        std::vector< std::shared_ptr<DirectionalSamplingS::Direction>> results;
-        for (std::shared_ptr<Direction> direction : directions)
-        {
-            if (function(direction))
-            {
-                results.push_back(direction);
-            }
-        }
-        return results;
-    }
 }
-
-
